Overhauled stake add
