+++ bittensor_cli/src/commands/stake/stake.py
@@ -1,18 +1,16 @@
 import asyncio
-from functools import partial
 
 from typing import TYPE_CHECKING, Optional
 import typer
 
 from bittensor_wallet import Wallet
 from bittensor_wallet.errors import KeyFileError
-from rich.prompt import Confirm, FloatPrompt, Prompt
+from rich.prompt import Confirm, Prompt
 from rich.table import Table
 from rich import box
 from rich.progress import Progress, BarColumn, TextColumn
 from rich.console import Group
 from rich.live import Live
-from async_substrate_interface.errors import SubstrateRequestException
 
 from bittensor_cli.src import COLOR_PALETTE
 from bittensor_cli.src.bittensor.balances import Balance
@@ -33,389 +31,6 @@ from bittensor_cli.src.bittensor.utils import (
 
 if TYPE_CHECKING:
     from bittensor_cli.src.bittensor.subtensor_interface import SubtensorInterface
-
-
-def _get_hotkeys_to_stake_to(
-    wallet: Wallet,
-    all_hotkeys: bool = False,
-    include_hotkeys: list[str] = None,
-    exclude_hotkeys: list[str] = None,
-) -> list[tuple[Optional[str], str]]:
-    """Get list of hotkeys to stake to based on input parameters.
-
-    Args:
-        wallet: The wallet containing hotkeys
-        all_hotkeys: If True, get all hotkeys from wallet except excluded ones
-        include_hotkeys: List of specific hotkeys to include (by name or ss58 address)
-        exclude_hotkeys: List of hotkeys to exclude when all_hotkeys is True
-
-    Returns:
-        List of tuples containing (hotkey_name, hotkey_ss58_address)
-        hotkey_name may be None if ss58 address was provided directly
-    """
-    if all_hotkeys:
-        # Stake to all hotkeys except excluded ones
-        all_hotkeys_: list[Wallet] = get_hotkey_wallets_for_wallet(wallet=wallet)
-        return [
-            (wallet.hotkey_str, wallet.hotkey.ss58_address)
-            for wallet in all_hotkeys_
-            if wallet.hotkey_str not in (exclude_hotkeys or [])
-        ]
-
-    if include_hotkeys:
-        print_verbose("Staking to only included hotkeys")
-        # Stake to specific hotkeys
-        hotkeys = []
-        for hotkey_ss58_or_hotkey_name in include_hotkeys:
-            if is_valid_ss58_address(hotkey_ss58_or_hotkey_name):
-                # If valid ss58 address, add directly
-                hotkeys.append((None, hotkey_ss58_or_hotkey_name))
-            else:
-                # If hotkey name, get ss58 from wallet
-                wallet_ = Wallet(
-                    path=wallet.path,
-                    name=wallet.name,
-                    hotkey=hotkey_ss58_or_hotkey_name,
-                )
-                hotkeys.append((wallet_.hotkey_str, wallet_.hotkey.ss58_address))
-        return hotkeys
-
-    # Default: stake to single hotkey from wallet
-    print_verbose(
-        f"Staking to hotkey: ({wallet.hotkey_str}) in wallet: ({wallet.name})"
-    )
-    assert wallet.hotkey is not None
-    return [(None, wallet.hotkey.ss58_address)]
-
-
-def _define_stake_add_table(wallet: Wallet, subtensor: "SubtensorInterface") -> Table:
-    """Creates and initializes a table for displaying stake information.
-
-    Args:
-        wallet: The wallet being used for staking
-        subtensor: The subtensor interface
-
-    Returns:
-        Table: An initialized rich Table object with appropriate columns
-    """
-    table = Table(
-        title=f"\n[{COLOR_PALETTE['GENERAL']['HEADER']}]Staking to:\n"
-        f"Wallet: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.name}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}], "
-        f"Coldkey ss58: [{COLOR_PALETTE['GENERAL']['COLDKEY']}]{wallet.coldkeypub.ss58_address}[/{COLOR_PALETTE['GENERAL']['COLDKEY']}]\n"
-        f"Network: {subtensor.network}[/{COLOR_PALETTE['GENERAL']['HEADER']}]\n",
-        show_footer=True,
-        show_edge=False,
-        header_style="bold white",
-        border_style="bright_black",
-        style="bold",
-        title_justify="center",
-        show_lines=False,
-        pad_edge=True,
-    )
-
-    table.add_column("Netuid", justify="center", style="grey89")
-    table.add_column(
-        "Hotkey", justify="center", style=COLOR_PALETTE["GENERAL"]["HOTKEY"]
-    )
-    table.add_column(
-        f"Amount ({Balance.get_unit(0)})",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["TAO"],
-    )
-    table.add_column(
-        f"Rate (per {Balance.get_unit(0)})",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["RATE"],
-    )
-    table.add_column(
-        "Received",
-        justify="center",
-        style=COLOR_PALETTE["POOLS"]["TAO_EQUIV"],
-    )
-    table.add_column(
-        "Slippage", justify="center", style=COLOR_PALETTE["STAKE"]["SLIPPAGE_PERCENT"]
-    )
-
-    return table
-
-
-def _print_table_and_slippage(table: Table, max_slippage: float):
-    """Prints the stake table, slippage warning, and table description.
-
-    Args:
-        table: The rich Table object to print
-        max_slippage: The maximum slippage percentage across all operations
-    """
-    console.print(table)
-
-    # Greater than 5%
-    if max_slippage > 5:
-        message = f"[{COLOR_PALETTE['STAKE']['SLIPPAGE_TEXT']}]-------------------------------------------------------------------------------------------------------------------\n"
-        message += f"[bold]WARNING:[/bold]  The slippage on one of your operations is high: [{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}]{max_slippage} %[/{COLOR_PALETTE['STAKE']['SLIPPAGE_PERCENT']}], this may result in a loss of funds.\n"
-        message += "-------------------------------------------------------------------------------------------------------------------\n"
-        console.print(message)
-
-    # Table description
-    console.print(
-        """
-[bold white]Description[/bold white]:
-The table displays information about the stake operation you are about to perform.
-The columns are as follows:
-    - [bold white]Netuid[/bold white]: The netuid of the subnet you are staking to.
-    - [bold white]Hotkey[/bold white]: The ss58 address of the hotkey you are staking to. 
-    - [bold white]Amount[/bold white]: The TAO you are staking into this subnet onto this hotkey.
-    - [bold white]Rate[/bold white]: The rate of exchange between your TAO and the subnet's stake.
-    - [bold white]Received[/bold white]: The amount of stake you will receive on this subnet after slippage.
-    - [bold white]Slippage[/bold white]: The slippage percentage of the stake operation. (0% if the subnet is not dynamic i.e. root).
-"""
-    )
-
-
-async def stake_add(
-    wallet: Wallet,
-    subtensor: "SubtensorInterface",
-    netuid: Optional[int],
-    stake_all: bool,
-    amount: float,
-    prompt: bool,
-    all_hotkeys: bool,
-    include_hotkeys: list[str],
-    exclude_hotkeys: list[str],
-):
-    """
-
-    Args:
-        wallet: wallet object
-        subtensor: SubtensorInterface object
-        netuid: the netuid to stake to (None indicates all subnets)
-        stake_all: whether to stake all available balance
-        amount: specified amount of balance to stake
-        delegate: whether to delegate stake, currently unused
-        prompt: whether to prompt the user
-        max_stake: maximum amount to stake (used in combination with stake_all), currently unused
-        all_hotkeys: whether to stake all hotkeys
-        include_hotkeys: list of hotkeys to include in staking process (if not specifying `--all`)
-        exclude_hotkeys: list of hotkeys to exclude in staking (if specifying `--all`)
-
-    Returns:
-
-    """
-    netuids = (
-        [int(netuid)]
-        if netuid is not None
-        else await subtensor.get_all_subnet_netuids()
-    )
-    hotkeys_to_stake_to = _get_hotkeys_to_stake_to(
-        wallet=wallet,
-        all_hotkeys=all_hotkeys,
-        include_hotkeys=include_hotkeys,
-        exclude_hotkeys=exclude_hotkeys,
-    )
-
-    # Determine the amount we are staking.
-    rows = []
-    stake_amount_balance = []
-    current_stake_balances = []
-    current_wallet_balance_ = await subtensor.get_balance(
-        wallet.coldkeypub.ss58_address
-    )
-    current_wallet_balance = current_wallet_balance_.set_unit(0)
-    remaining_wallet_balance = current_wallet_balance
-    max_slippage = 0.0
-
-    starting_chain_head = await subtensor.substrate.get_chain_head()
-    _all_dynamic_info, stake_info_dict = await asyncio.gather(
-        subtensor.all_subnets(),
-        subtensor.get_stake_for_coldkey(
-            coldkey_ss58=wallet.coldkeypub.ss58_address,
-            block_hash=starting_chain_head,
-        ),
-    )
-    all_dynamic_info = {di.netuid: di for di in _all_dynamic_info}
-    initial_stake_balances = {}
-    for _, hotkey_ss58 in hotkeys_to_stake_to:
-        initial_stake_balances[hotkey_ss58] = {}
-        for netuid in netuids:
-            initial_stake_balances[hotkey_ss58][netuid] = Balance.from_rao(0)
-
-    for stake_info in stake_info_dict:
-        if stake_info.hotkey_ss58 in initial_stake_balances:
-            initial_stake_balances[stake_info.hotkey_ss58][stake_info.netuid] = (
-                stake_info.stake
-            )
-
-    for hk_name, hk_ss58 in hotkeys_to_stake_to:
-        if not is_valid_ss58_address(hk_ss58):
-            print_error(
-                f"The entered hotkey ss58 address is incorrect: {hk_name} | {hk_ss58}"
-            )
-            return False
-    for hotkey in hotkeys_to_stake_to:
-        for netuid in netuids:
-            # Check that the subnet exists.
-            dynamic_info = all_dynamic_info.get(netuid)
-            if not dynamic_info:
-                err_console.print(f"Subnet with netuid: {netuid} does not exist.")
-                continue
-            current_stake_balances.append(initial_stake_balances[hotkey[1]][netuid])
-
-            # Get the amount.
-            amount_to_stake_as_balance = Balance(0)
-            if amount:
-                amount_to_stake_as_balance = Balance.from_tao(amount)
-            elif stake_all:
-                amount_to_stake_as_balance = current_wallet_balance / len(netuids)
-            elif not amount:
-                if Confirm.ask(f"Stake all: [bold]{remaining_wallet_balance}[/bold]?"):
-                    amount_to_stake_as_balance = remaining_wallet_balance
-                else:
-                    try:
-                        amount = FloatPrompt.ask(
-                            f"Enter amount to stake in {Balance.get_unit(0)} to subnet: {netuid}"
-                        )
-                        amount_to_stake_as_balance = Balance.from_tao(amount)
-                    except ValueError:
-                        err_console.print(
-                            f":cross_mark:[red]Invalid amount: {amount}[/red]"
-                        )
-                        return False
-            stake_amount_balance.append(amount_to_stake_as_balance)
-
-            # Check enough to stake.
-            amount_to_stake_as_balance.set_unit(0)
-            if amount_to_stake_as_balance > remaining_wallet_balance:
-                err_console.print(
-                    f"[red]Not enough stake[/red]:[bold white]\n wallet balance:{remaining_wallet_balance} < "
-                    f"staking amount: {amount_to_stake_as_balance}[/bold white]"
-                )
-                return False
-            remaining_wallet_balance -= amount_to_stake_as_balance
-
-            # Slippage warning
-            received_amount, _, slippage_pct_float = (
-                dynamic_info.tao_to_alpha_with_slippage(amount_to_stake_as_balance)
-            )
-            if dynamic_info.is_dynamic:
-                slippage_pct = f"{slippage_pct_float:.4f} %"
-                rate = str(1 / (float(dynamic_info.price) or 1))
-            else:
-                slippage_pct_float = 0
-                slippage_pct = f"[{COLOR_PALETTE['STAKE']['SLIPPAGE_TEXT']}]N/A[/{COLOR_PALETTE['STAKE']['SLIPPAGE_TEXT']}]"
-                rate = str(1)
-            max_slippage = max(slippage_pct_float, max_slippage)
-            rows.append(
-                (
-                    str(netuid),
-                    # f"{staking_address_ss58[:3]}...{staking_address_ss58[-3:]}",
-                    f"{hotkey[1]}",
-                    str(amount_to_stake_as_balance),
-                    rate + f" {Balance.get_unit(netuid)}/{Balance.get_unit(0)} ",
-                    str(received_amount.set_unit(netuid)),
-                    str(slippage_pct),
-                )
-            )
-
-    # Define and print stake table + slippage warning
-    table = _define_stake_add_table(wallet, subtensor)
-    for row in rows:
-        table.add_row(*row)
-    _print_table_and_slippage(table, max_slippage)
-
-    if prompt:
-        if not Confirm.ask("Would you like to continue?"):
-            raise typer.Exit()
-
-    async def send_extrinsic(
-        netuid_i, amount_, current, staking_address_ss58, status=None
-    ):
-        err_out = partial(print_error, status=status)
-        failure_prelude = (
-            f":cross_mark: [red]Failed[/red] to stake {amount} on Netuid {netuid_i}"
-        )
-        call = await subtensor.substrate.compose_call(
-            call_module="SubtensorModule",
-            call_function="add_stake",
-            call_params={
-                "hotkey": staking_address_ss58,
-                "netuid": netuid_i,
-                "amount_staked": amount_.rao,
-            },
-        )
-        extrinsic = await subtensor.substrate.create_signed_extrinsic(
-            call=call, keypair=wallet.coldkey
-        )
-        try:
-            response = await subtensor.substrate.submit_extrinsic(
-                extrinsic, wait_for_inclusion=True, wait_for_finalization=False
-            )
-        except SubstrateRequestException as e:
-            err_out(
-                f"\n{failure_prelude} with error: {format_error_message(e, subtensor.substrate)}"
-            )
-            return
-        else:
-            await response.process_events()
-            if not await response.is_success:
-                err_out(
-                    f"\n{failure_prelude} with error: {format_error_message(await response.error_message, subtensor.substrate)}"
-                )
-            else:
-                new_balance, stake_info_dict = await asyncio.gather(
-                    subtensor.get_balance(wallet.coldkeypub.ss58_address),
-                    subtensor.get_stake_for_coldkey(
-                        coldkey_ss58=wallet.coldkeypub.ss58_address,
-                    ),
-                )
-                new_stake = Balance.from_rao(0)
-                for stake_info in stake_info_dict:
-                    if (
-                        stake_info.hotkey_ss58 == staking_address_ss58
-                        and stake_info.netuid == netuid_i
-                    ):
-                        new_stake = stake_info.stake.set_unit(netuid_i)
-                        break
-                console.print(":white_heavy_check_mark: [green]Finalized[/green]")
-                console.print(
-                    f"Balance:\n  [blue]{current_wallet_balance}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_balance}"
-                )
-                console.print(
-                    f"Subnet: [{COLOR_PALETTE['GENERAL']['SUBHEADING']}]{netuid_i}[/{COLOR_PALETTE['GENERAL']['SUBHEADING']}] Stake:\n  [blue]{current}[/blue] :arrow_right: [{COLOR_PALETTE['STAKE']['STAKE_AMOUNT']}]{new_stake}"
-                )
-
-    # Perform staking operation.
-    try:
-        wallet.unlock_coldkey()
-    except KeyFileError:
-        err_console.print("Error decrypting coldkey (possibly incorrect password)")
-        return False
-    extrinsics_coroutines = [
-        send_extrinsic(ni, am, curr, staking_address)
-        for i, (ni, am, curr) in enumerate(
-            zip(netuids, stake_amount_balance, current_stake_balances)
-        )
-        for _, staking_address in hotkeys_to_stake_to
-    ]
-    if len(extrinsics_coroutines) == 1:
-        with console.status(f"\n:satellite: Staking on netuid(s): {netuids} ..."):
-            await extrinsics_coroutines[0]
-    else:
-        with console.status(":satellite: Checking transaction rate limit ..."):
-            tx_rate_limit_blocks = await subtensor.query(
-                module="SubtensorModule", storage_function="TxRateLimit"
-            )
-        netuid_hk_pairs = [(ni, hk) for ni in netuids for hk in hotkeys_to_stake_to]
-        for item, kp in zip(extrinsics_coroutines, netuid_hk_pairs):
-            ni, hk = kp
-            with console.status(
-                f"\n:satellite: Staking on netuid {ni} with hotkey {hk}... ..."
-            ):
-                await item
-            if tx_rate_limit_blocks > 0:
-                with console.status(
-                    f":hourglass: [yellow]Waiting for tx rate limit:"
-                    f" [white]{tx_rate_limit_blocks}[/white] blocks[/yellow]"
-                ):
-                    await asyncio.sleep(tx_rate_limit_blocks * 12)  # 12 sec per block
 
 
 async def unstake_selection(
